+++ kitty/glfw-wrapper.h
@@ -2181,10 +2181,6 @@ typedef void* (*glfwGetCocoaWindow_func)(GLFWwindow*);
 GFW_EXTERN glfwGetCocoaWindow_func glfwGetCocoaWindow_impl;
 #define glfwGetCocoaWindow glfwGetCocoaWindow_impl
 
-typedef void (*glfwHideCocoaTitlebar_func)(GLFWwindow*, bool);
-GFW_EXTERN glfwHideCocoaTitlebar_func glfwHideCocoaTitlebar_impl;
-#define glfwHideCocoaTitlebar glfwHideCocoaTitlebar_impl
-
 typedef void* (*glfwGetNSGLContext_func)(GLFWwindow*);
 GFW_EXTERN glfwGetNSGLContext_func glfwGetNSGLContext_impl;
 #define glfwGetNSGLContext glfwGetNSGLContext_impl
@@ -2240,6 +2236,10 @@ GFW_EXTERN glfwGetX11Window_func glfwGetX11Window_impl;
 typedef void (*glfwSetPrimarySelectionString_func)(GLFWwindow*, const char*);
 GFW_EXTERN glfwSetPrimarySelectionString_func glfwSetPrimarySelectionString_impl;
 #define glfwSetPrimarySelectionString glfwSetPrimarySelectionString_impl
+
+typedef void (*glfwCocoaSetWindowChrome_func)(GLFWwindow*, unsigned int, bool, unsigned int, int, unsigned int, bool, int, float, bool);
+GFW_EXTERN glfwCocoaSetWindowChrome_func glfwCocoaSetWindowChrome_impl;
+#define glfwCocoaSetWindowChrome glfwCocoaSetWindowChrome_impl
 
 typedef const char* (*glfwGetPrimarySelectionString_func)(GLFWwindow*);
 GFW_EXTERN glfwGetPrimarySelectionString_func glfwGetPrimarySelectionString_impl;